+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.9] - 2024-10-05
+
+### Added
+
+- Add option to mimic Graphviz Subgraph
+  - Get-DiaHTMLNodeTable
+  - Get-DiaHTMLTable
+  - Get-DiaHtmlSubGraph
+- Add cmdlet to create the Signature table
+  - Get-DiaHtmlSignatureTable
+
+### Changed
+
+- Update Graphviz binaries to v12.1.2
+- Update modules tests
+
 ## [0.2.8] - 2024-09-17
 
 ### Added